% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manageDic.R
\name{buildDictionary}
\alias{buildDictionary}
\title{buildDictionary}
\usage{
buildDictionary(ext_dic = "sejong", category_dic_nms = "",
  user_dic = data.frame(), replace_usr_dic = F)
}
\arguments{
\item{ext_dic}{}

\item{user_dic}{}

\item{replace_usr_dic}{}

\item{category_dic_ids}{}
}
\description{
buildDictionary
}

