\name{buildUserDic}
\alias{buildUserDic}
\title{"dic_usr.txt" editor function}
\usage{
  buildUserDic(newUserDic, append = TRUE, verbose = FALSE)
}
\arguments{
  \item{newUserDic}{new user dictionary as data.frame}

  \item{append}{append or replacing}

  \item{verbose}{see detail error logs}
}
\description{
  merging or replacing current dic_user.txt file.
}

