% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Concordances.R
\name{concordance_str}
\alias{concordance_str}
\title{concordance for input text vector}
\usage{
concordance_str(string, pattern, span = 5)
}
\arguments{
\item{string}{input text as character vector or single character}

\item{pattern}{patterns of central words}

\item{span}{how many character will be produced around input pattern}
}
\description{
returns concordance text for input pattern and span.
}
\author{
Heewon Jeon
}
\references{
Church, K. W. and Mercer, R. L. (1993). Introduction to the special issue on computational linguistics using large corpora. Computational Linguistics, 19(1):1-24.
}

