\name{doKoMorph}
\alias{doKoMorph}
\title{doKoMorph...}
\usage{doKoMorph(sentence)}
\description{doKoMorph}
\details{morphlogical analyze the sentence.
it uses lucene korean analyzer.
see details in \url{http://sourceforge.net/projects/lucenekorean/}}
\value{stem of sentence}
\arguments{\item{sentence}{input}}
