% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manageDic.R
\name{get_dictionary}
\alias{get_dictionary}
\title{Get Dictionary}
\usage{
get_dictionary(dic_name)
}
\arguments{
\item{dic_name}{one of dictionary name(character), \strong{woorimalsam}, \strong{insighter}, \strong{sejong}, \strong{user_dic}}
}
\value{
The \code{data.frame} object contains tags and terms
}
\description{
Get Dictionary
}
\examples{
\dontrun{
   dic_df <- get_dictionary('sejong')
}
}

