% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hangulUtils.R
\name{is.ascii}
\alias{is.ascii}
\title{check if sentence is all ASCII}
\usage{
is.ascii(sentence)
}
\arguments{
\item{sentence}{input charactors}
}
\value{
TRUE or FALSE
}
\description{
Function checks with each charactor is ASCII
}

