\name{is.hangul}
\alias{is.hangul}
\title{check if sentence is all Hangul.}
\usage{
  is.hangul(sentenceU8)
}
\arguments{
  \item{sentenceU8}{input charactors(must be UTF-8)}
}
\value{
  TRUE or FALSE
}
\description{
  Function checks if each charactor is Hangul or Jamo.
  Example will be shown in
  \href{https://github.com/haven-jeon/KoNLP/wiki}{github
  wiki}.
}

