% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hangulUtils.R
\name{is.jaeum}
\alias{is.jaeum}
\title{check if sentence is all Jaeum}
\usage{
is.jaeum(sentence)
}
\arguments{
\item{sentence}{input charactors}
}
\value{
TRUE or FALSE
}
\description{
Function checks with each charactor is Jaeum
}

