\name{is.jamo}
\alias{is.jamo}
\title{check if sentence is all Jamo}
\usage{
  is.jamo(sentenceU8)
}
\arguments{
  \item{sentenceU8}{input charactors(must be UTF-8)}
}
\value{
  TRUE or FALSE
}
\description{
  Function checks with each charactor is Jamo. Example will
  be shown in
  \href{https://github.com/haven-jeon/KoNLP/wiki}{github
  wiki}.
}

