\name{is.moeum}
\alias{is.moeum}
\title{check if sentence is all Moeum}
\usage{
  is.moeum(sentenceU8)
}
\arguments{
  \item{sentenceU8}{input charactors(must be UTF-8)}
}
\value{
  TRUE or FALSE
}
\description{
  Function checks with each charactor is Moeum Example will
  be shown in
  \href{https://github.com/haven-jeon/KoNLP/wiki}{github
  wiki}.
}

