\name{mutualinformation}
\alias{mutualinformation}
\title{mutual information for input text}
\usage{
  mutualinformation(text, query = "",
    method = c("mutual", "tscores"))
}
\arguments{
  \item{text}{input character vector}

  \item{method}{for calculations(`mutual' or `t-scores')}
}
\description{
  returns mutual information or t-scores for input text
}
\author{
  Heewon Jeon
}

