% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manageDic.R
\name{reloadUserDic}
\alias{reloadUserDic}
\title{reload dictionaries for specific functions}
\usage{
reloadUserDic(whichDics)
}
\arguments{
\item{whichDics}{character vector which can be "extractNoun", "SimplePos09", "SimplePos22", "SimplePos22"}
}
\description{
This function for reloading user dictionary for specific functions,
after you have updated user dictionary on KoNLP_dic/current/user_dic.txt.
}
\examples{
\dontrun{
reloadUserDic(c("extractNoun", "SimplePos22"))} 
}

