% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scala_library_install}
\alias{scala_library_install}
\title{scala_library_install}
\usage{
scala_library_install(ver = "2.11.8")
}
\arguments{
\item{ver}{which scala version to install}
}
\description{
scala_library_install
}

