% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koAnalyzerRun.R
\name{SimplePos09}
\alias{SimplePos09}
\title{POS tagging by using 9 KAIST tags}
\usage{
SimplePos09(sentence, autoSpacing = FALSE)
}
\arguments{
\item{sentence}{input}

\item{autoSpacing}{boolean dees it need to apply auto-spacing for input. defaul\code{FALSE}}
}
\value{
results of tagged analysis
}
\description{
Do pos tagging using 9 tags uses Hannanum analyzer.
see details in \href{http://semanticweb.kaist.ac.kr/home/index.php/HanNanum}{Hannanum}. 
Example will be shown in \href{https://github.com/haven-jeon/KoNLP/wiki}{github wiki}.
}
\references{
Sangwon Park et al(2010). A Plug-In Component-based Korean Morphological Analyzer
}

