% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manageDic.R
\name{useInsighterDic}
\alias{useInsighterDic}
\title{use Insighter dictionary}
\usage{
useInsighterDic(backup = T)
}
\arguments{
\item{backup}{will backup current working dictionary?}
}
\description{
use Insighter dictionary
}

