% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manageDic.R
\name{useNIADic}
\alias{useNIADic}
\title{use Insighter and Woorimalsam dictionary}
\usage{
useNIADic(backup = T)
}
\arguments{
\item{backup}{will backup current working dictionary?}
}
\description{
use Insighter and Woorimalsam dictionary
}

