% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manageDic.R
\name{useSejongDic}
\alias{useSejongDic}
\title{use Sejong noun dictionary}
\usage{
useSejongDic(backup = T)
}
\arguments{
\item{backup}{will backup current dictionary?}
}
\description{
Retrive Sejong dictionary to use in KoNLP
}
\references{
\url{http://www.sejong.or.kr/}
}

