% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manageDic.R
\name{useSystemDic}
\alias{useSystemDic}
\title{use system default dictionary}
\usage{
useSystemDic(backup = T)
}
\arguments{
\item{backup}{will backup current dictionary?}
}
\description{
Retrive system default dictionary to use in KoNLP
}

