% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manageDic.R
\name{useWoorimalsamDic}
\alias{useWoorimalsamDic}
\title{use Woorimalsam dictionary}
\usage{
useWoorimalsamDic(backup = T)
}
\arguments{
\item{backup}{will backup current working dictionary?}
}
\description{
use Woorimalsam dictionary
}

