% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manageDic.R
\name{backupUsrDic}
\alias{backupUsrDic}
\title{use for backup current dic_user.txt
 
Utility function for backup dic_user.txt file to backup directory.}
\usage{
backupUsrDic(ask = TRUE)
}
\arguments{
\item{ask}{ask to confirm backup}
}
\description{
use for backup current dic_user.txt
 
Utility function for backup dic_user.txt file to backup directory.
}
\examples{
\dontrun{
## This codes can not be run if you don't have encoding system 
## which can en/decode Hangul(ex) CP949, EUC-KR, UTF-8). 
dicpath <- file.path(system.file(package="Sejong"), "dics", "handic.zip")
conn <- unz(dicpath, file.path("data","kE","dic_user2.txt"))
newdic <- read.csv(conn, sep="\\t", header=FALSE, fileEncoding="UTF-8", stringsAsFactors=FALSE)
mergeUserDic(newdic)
## backup merged new dictionary
backupUsrDic(ask=FALSE)
## restore from backup directory
restoreUsrDic(ask=FALSE)
## reloading new dictionary
reloadAllDic()}
}

