\name{convertHangulStringToJamos}
\alias{convertHangulStringToJamos}
\title{convertHangulStringToJamos...}
\usage{convertHangulStringToJamos(hangul)}
\description{convertHangulStringToJamos}
\details{convert Hangul sentence to Jamos(now on testing~)}
\value{Jamo sequences}
\arguments{\item{hangul}{hangul string}}
