% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hangulUtils.R
\name{convertHangulStringToJamos}
\alias{convertHangulStringToJamos}
\title{convertion function  Hangul string to Jamos}
\usage{
convertHangulStringToJamos(hangul)
}
\arguments{
\item{hangul}{Hangul string}
}
\value{
Jamo sequences
}
\description{
convert Hangul sentence to Jamos.
Example will be shown in \href{https://github.com/haven-jeon/KoNLP/wiki}{github wiki}.
}

