\name{convertHangulStringToKeyStrokes}
\alias{convertHangulStringToKeyStrokes}
\title{convertHangulStringToKeyStrokes...}
\usage{convertHangulStringToKeyStrokes(hangul)}
\description{convertHangulStringToKeyStrokes}
\details{convert Hangul String to Keystrokes, each Hangul syllable can be dilimitered by \emph{OxFF5C}.}
\value{Keystroke sequence}
\arguments{\item{hangul}{hangul sentence}}
