\name{convertHangulStringToKeyStrokes}
\alias{convertHangulStringToKeyStrokes}
\title{convertHangulStringToKeyStrokes}
\usage{
  convertHangulStringToKeyStrokes(hangul)
}
\arguments{
  \item{hangul}{hangul sentence}
}
\value{
  Keystroke sequence
}
\description{
  convert Hangul String to Keystrokes, each Hangul syllable
  can be dilimitered by \emph{OxFF5C}.
}

