% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hangulUtils.R
\name{convertHangulStringToKeyStrokes}
\alias{convertHangulStringToKeyStrokes}
\title{convertion function Hangul string to keyStrokes}
\usage{
convertHangulStringToKeyStrokes(hangul, isFullwidth = TRUE)
}
\arguments{
\item{hangul}{Hangul sentence}

\item{isFullwidth}{specify returned character will be Fullwidth ASCII or Halfwidth ASCII}
}
\value{
Keystroke sequence
}
\description{
Function can convert Hangul string to Keystrokes. 
Example will be shown in \href{https://github.com/haven-jeon/KoNLP/wiki}{github wiki}.
}

