\name{detectInputEncoding}
\alias{detectInputEncoding}
\title{Rough encoding detection function}
\usage{
  detectInputEncoding(charinput)
}
\arguments{
  \item{charinput}{charvector}
}
\value{
  encoding names of rawinpus.
}
\description{
  function to be used for file or raw vector encodoing
  detection. This is for internal use.
}

