\name{doKoMorph}
\alias{doKoMorph}
\title{doKoMorph}
\usage{
  doKoMorph(sentence)
}
\arguments{
  \item{sentence}{input}
}
\value{
  stem of sentence
}
\description{
  morphlogical analyze the sentence. it uses lucene korean
  analyzer. see details in
  \url{http://sourceforge.net/projects/lucenekorean/}
}

