\name{doKoMorph}
\alias{doKoMorph}
\title{morphlogical analysis function for Hangul}
\usage{
  doKoMorph(sentence)
}
\arguments{
  \item{sentence}{input}
}
\value{
  stem of sentence
}
\description{
  morphlogical analyze the sentence uses lucene korean
  analyzer. Example will be shown in
  \href{https://github.com/haven-jeon/KoNLP/wiki}{github
  wiki}.
}

