\name{extractNoun}
\alias{extractNoun}
\title{extractNoun}
\usage{
  extractNoun(sentence)
}
\arguments{
  \item{sentence}{input}
}
\value{
  Noun of sentence
}
\description{
  extract Nouns from Korean sentence. it uses Hannanum
  analyzer. see detail in
  \url{http://semanticweb.kaist.ac.kr/home/index.php/HanNanum}
}

