\name{extractNoun}
\alias{extractNoun}
\title{Noun extractor for Hangul}
\usage{
  extractNoun(sentence)
}
\arguments{
  \item{sentence}{input}
}
\value{
  Noun of sentence
}
\description{
  extract Nouns from Korean sentence uses Hannanum
  analyzer. see detail in
  \href{http://semanticweb.kaist.ac.kr/home/index.php/HanNanum}{Hannanum}.
  Example will be shown in
  \href{https://github.com/haven-jeon/KoNLP/wiki}{github
  wiki}.
}

