\name{extractNoun}
\alias{extractNoun}
\title{extractNoun...}
\usage{extractNoun(sentence)}
\description{extractNoun}
\details{extract Nouns from Korean sentence. 
it uses Hannanum analyzer.
see detail in \url{http://semanticweb.kaist.ac.kr/home/index.php/HanNanum}}
\value{Noun of sentence}
\arguments{\item{sentence}{input}}
