% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koAnalyzerRun.R
\name{extractNoun}
\alias{extractNoun}
\title{Noun extractor for Hangul}
\usage{
extractNoun(sentences, autoSpacing = FALSE)
}
\arguments{
\item{sentences}{input character vector}

\item{autoSpacing}{boolean dees it need to apply auto-spacing for input. defaul\code{FALSE}}
}
\value{
Nouns of sentences, returns \code{list} if input is character vector of more than 2 sentences.
}
\description{
extract Nouns from Korean sentence uses Hannanum analyzer.
see detail in \href{http://semanticweb.kaist.ac.kr/home/index.php/HanNanum}{Hannanum}. 
Example will be shown in \href{https://github.com/haven-jeon/KoNLP/wiki}{github wiki}.
}
\references{
Sangwon Park et al(2010). A Plug-In Component-based Korean Morphological Analyzer
}

