\name{is.hangul}
\alias{is.hangul}
\title{is.hangul}
\usage{
  is.hangul(sentence)
}
\arguments{
  \item{sentence}{input charactor}
}
\value{
  TRUE or FALSE of sentence vector(s)
}
\description{
  checking sentence is hangul or not. Example will be shown
  in \href{https://github.com/haven-jeon/KoNLP/wiki}{github
  wiki}.
}

