% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hangulUtils.R
\name{is.hangul}
\alias{is.hangul}
\title{check if sentence is all Hangul}
\usage{
is.hangul(sentence)
}
\arguments{
\item{sentence}{input charactors}
}
\value{
TRUE or FALSE
}
\description{
Function checks if each charactor is Hangul or Jamo. 
Example will be shown in \href{https://github.com/haven-jeon/KoNLP/wiki}{github wiki}.
}

