% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hangulUtils.R
\name{is.jamo}
\alias{is.jamo}
\title{check if sentence is all Jamo}
\usage{
is.jamo(sentence)
}
\arguments{
\item{sentence}{input charactors}
}
\value{
TRUE or FALSE
}
\description{
Function checks with each charactor is Jamo. 
Example will be shown in \href{https://github.com/haven-jeon/KoNLP/wiki}{github wiki}.
}

