% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hangulUtils.R
\name{is.moeum}
\alias{is.moeum}
\title{check if sentence is all Moeum}
\usage{
is.moeum(sentence)
}
\arguments{
\item{sentence}{input charactors}
}
\value{
TRUE or FALSE
}
\description{
Function checks with each charactor is Moeum
}

