\name{MorphAnalyzer}
\alias{MorphAnalyzer}
\title{MorphAnalyzer}
\usage{
  MorphAnalyzer(sentence)
}
\arguments{
  \item{sentence}{input}
}
\value{
  result of analysis
}
\description{
  Do the morphological analysis, not doing pos tagging. it
  uses Hannanum analyzer. see details in
  \url{http://semanticweb.kaist.ac.kr/home/index.php/HanNanum}
}

