\name{MorphAnalyzer}
\alias{MorphAnalyzer}
\title{MorphAnalyzer...}
\usage{MorphAnalyzer(sentence)}
\description{MorphAnalyzer}
\details{Do the morphological analysis, not doing pos tagging. 
it uses Hannanum analyzer.
see details in \url{http://semanticweb.kaist.ac.kr/home/index.php/HanNanum}}
\value{result of analysis}
\arguments{\item{sentence}{input}}
