\name{MorphAnalyzer}
\alias{MorphAnalyzer}
\title{Hannanum morphological analyzer interface function}
\usage{
  MorphAnalyzer(sentence)
}
\arguments{
  \item{sentence}{input}
}
\value{
  result of analysis
}
\description{
  Do the morphological analysis, not doing pos tagging uses
  Hannanum analyzer. see details in
  \href{http://semanticweb.kaist.ac.kr/home/index.php/HanNanum}{Hannanum}.
  Example will be shown in
  \href{https://github.com/haven-jeon/KoNLP/wiki}{github
  wiki}.
}

