% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koAnalyzerRun.R
\name{MorphAnalyzer}
\alias{MorphAnalyzer}
\title{Hannanum morphological analyzer interface function}
\usage{
MorphAnalyzer(sentences, autoSpacing = FALSE)
}
\arguments{
\item{sentences}{input character vector}

\item{autoSpacing}{boolean dees it need to apply auto-spacing for input. defaul\code{FALSE}}
}
\value{
morphemes of sentences
}
\description{
Do the morphological analysis, not doing pos tagging uses Hannanum analyzer.
see details in \href{http://semanticweb.kaist.ac.kr/home/index.php/HanNanum}{Hannanum}. 
Example will be shown in \href{https://github.com/haven-jeon/KoNLP/wiki}{github wiki}.
}
\references{
Sangwon Park et al(2010). A Plug-In Component-based Korean Morphological Analyzer
}

