\name{reloadAllDic}
\alias{reloadAllDic}
\title{reload all Hannanum analyzer dictionary}
\usage{
  reloadAllDic()
}
\description{
  Mainly, user dictionary reloading for Hannanum Analyzer.
  If you want to update user dictionary on
  KoNLP_pkg_dir/inst/dics/data/kE/dic_user.txt, need to
  execute this function after editing dic.
}

