\name{SimplePos09}
\alias{SimplePos09}
\title{SimplePos09...}
\usage{SimplePos09(sentence)}
\description{SimplePos09}
\details{Do pos tagging using 9 tags. 
it uses Hannanum analyzer.
see details in \url{http://semanticweb.kaist.ac.kr/home/index.php/HanNanum}}
\value{Noun of sentence}
\arguments{\item{sentence}{input}}
