\name{SimplePos09}
\alias{SimplePos09}
\title{SimplePos09}
\usage{
  SimplePos09(sentence)
}
\arguments{
  \item{sentence}{input}
}
\value{
  Noun of sentence
}
\description{
  Do pos tagging using 9 tags. it uses Hannanum analyzer.
  see details in
  \url{http://semanticweb.kaist.ac.kr/home/index.php/HanNanum}
}

