\name{SimplePos09}
\alias{SimplePos09}
\title{POS tagging by using 9 KAIST tags}
\usage{
  SimplePos09(sentence)
}
\arguments{
  \item{sentence}{input}
}
\value{
  results of tagged analysis
}
\description{
  Do pos tagging using 9 tags uses Hannanum analyzer. see
  details in
  \href{http://semanticweb.kaist.ac.kr/home/index.php/HanNanum}{Hannanum}.
  Example will be shown in
  \href{https://github.com/haven-jeon/KoNLP/wiki}{github
  wiki}.
}

