\name{SimplePos22}
\alias{SimplePos22}
\title{SimplePos22}
\usage{
  SimplePos22(sentence)
}
\arguments{
  \item{sentence}{input}
}
\value{
  result of analysis
}
\description{
  Do pos tagging using 22 tags uses Hannanum analyzer. see
  details in
  \href{http://semanticweb.kaist.ac.kr/home/index.php/HanNanum}{Hannanum}.
  Example will be shown in
  \href{https://github.com/haven-jeon/KoNLP/wiki}{github
  wiki}.
}

