\name{SimplePos22}
\alias{SimplePos22}
\title{SimplePos22...}
\usage{SimplePos22(sentence)}
\description{SimplePos22}
\details{Do pos tagging using 22 tags. 
it uses Hannanum analyzer.
see details in \url{http://semanticweb.kaist.ac.kr/home/index.php/HanNanum}}
\value{result of tags}
\arguments{\item{sentence}{input}}
